refactor(admin): split monolithic admin.module.ts into per-concern files (#1907)
Extract AdminService and AdminController out of the 88-line
admin.module.ts into admin.service.ts and admin.controller.ts, matching
the file-per-concern layout used by the other modules. AdminModule now
only wires the pieces together. No behavior change; existing admin unit
tests still pass.
